# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/oauth/authorize": {
      "get": {
        "operationId": "AuthorizationController_authorize",
        "summary": "Authorize for access_code",
        "parameters": [
          {
            "name": "response_type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "access_code"
          }
        },
        "tags": [
          "Authorization"
        ]
      }
    },
...
  },
  "info": {
    "title": "AIO_API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}